o Fixed incompatibility with Java 1.3.
git-svn-id: http://svn.code.sf.net/p/jtds/code/branches/jTDS_1_2_2_JDBC3/jtds@1041 97a8069c-bbaa-4fa2-9e49-561d8e0b19ef

+++ src/test/net/sourceforge/jtds/test/PreparedStatementTest.java
@@ -22,7 +22,7 @@ import java.sql.*;
 import java.util.*;
 
 /**
- * @version $Id: PreparedStatementTest.java,v 1.46.2.2 2009-08-21 15:42:17 ickzon Exp $
+ * @version $Id: PreparedStatementTest.java,v 1.46.2.3 2009-08-23 10:50:43 ickzon Exp $
  */
 public class PreparedStatementTest extends TestBase {
 
@@ -1104,7 +1104,7 @@ public class PreparedStatementTest extends TestBase {
         long seed = System.currentTimeMillis();
         Random r = new Random(seed);
 
-        Float value = 0.000803f;
+        Float value = Float.valueOf(0.000803f);
 
         PreparedStatement ps1 = con.prepareStatement("update #testArithemicOverflow set data = ? where id = ?");
         PreparedStatement ps2 = con.prepareStatement("select data from #testArithemicOverflow where id = ?");
@@ -1112,17 +1112,17 @@ public class PreparedStatementTest extends TestBase {
         try {
             for (int i = 0; i < 1000; i++) {
                 if (i > 0) {
-                    value = r.nextFloat() * Float.MAX_VALUE * (r.nextBoolean() ? 1 : -1);
+                    value = Float.valueOf(r.nextFloat() * Float.MAX_VALUE * (r.nextBoolean() ? 1 : -1));
                 }
 
-                ps1.setFloat(1, value);
+                ps1.setFloat(1, value.floatValue());
                 ps1.setInt(2, 1);
                 assertEquals(1, ps1.executeUpdate());
 
                 ps2.setInt(1, 1);
                 ResultSet rs = ps2.executeQuery();
                 assertTrue(rs.next());
-                assertEquals(value, rs.getFloat(1));
+                assertEquals(value, Float.valueOf(rs.getFloat(1)));
                 rs.close();
             }
         } catch (Throwable t) {
